代码优化，加入log文件夹

+++ anjukespider/pipelines.py
@@ -1,4 +1,6 @@
 import json
+import os
+import logging
 
 from scrapy.pipelines.files import FilesPipeline
 from anjukespider.model.scene import Scene
@@ -6,19 +8,20 @@ from anjukespider.model.config import DBSession
 from anjukespider.model.config import Redis
 from anjukespider.items import AnjukespiderItem
 from anjukespider.items import FileItem
+from anjukespider.settings import FILES_STORE
 from scrapy.exceptions import DropItem
 from scrapy import Request
 
 
 # 数据去重
-class DuplicatesPipeline(object):
-    def process_item(self, item, spider):
-        if Redis.hexists("duplicate", item['scene_name']):
-            print("Duplicate item found: %s" % item)
-            raise DropItem("Duplicate item found: %s" % item)
-        else:
-            Redis.hset("duplicate", item['scene_name'], 1)
-            return item
+# class DuplicatesPipeline(object):
+#     def process_item(self, item, spider):
+#         if Redis.hexists("duplicate", item['scene_name']):
+#             print("Duplicate item found: %s" % item)
+#             raise DropItem("Duplicate item found: %s" % item)
+#         else:
+#             Redis.hset("duplicate", item['scene_name'], 1)
+#             return item
 
 
 # 下载图片（自带）
@@ -31,16 +34,20 @@ class ImgDownloadPipeline(FilesPipeline):
 
     def get_media_requests(self, item, info):
         if isinstance(item, FileItem):
+            # logging.info("下载图片...")
+            print("下载图片...")
             for file_url in item['file_urls']:
                 yield Request(url=file_url, meta={'item': item})
 
     def item_completed(self, results, item, info):
         if isinstance(item, FileItem):
             file_paths = [x['path'] for ok, x in results if ok]
+            # logging.info("file_paths：", file_paths)
             print("file_paths：", file_paths)
             if not file_paths:
                 raise DropItem("Item contains no images")
             item['file_paths'] = file_paths
+            print("hotspots_index：", item["hotspots_index"])
             return item
         else:
             return item
@@ -50,8 +57,12 @@ class ImgDownloadPipeline(FilesPipeline):
 class JsonDownloadPipeline(object):
     def process_item(self, item, spider):
         if isinstance(item, FileItem):
-            with open("F:\\Project\\Scrapy_anjuke\\anjukespider\\scene\\%s\\scene.json" % item["file_name"], 'w') as file:
-                json.dump(item["file_json"], file)
+            scene_path = FILES_STORE + item["file_name"]
+            if not os.path.exists(scene_path + "\\scene.json"):
+                # logging.info("下载Json文件")
+                print("下载Json文件")
+                with open(scene_path + "\\scene.json", 'w') as file:
+                    json.dump(item["file_json"], file)
         return item
 
 
@@ -63,16 +74,20 @@ class DBPipeline(object):
     def process_item(self, item, spider):
         if isinstance(item, AnjukespiderItem):
             try:
+                # logging.info("录入数据库")
+                print("录入数据库")
                 sql = Scene(
                     scene_name=item["scene_name"],
                     web_site=item["web_site"],
                     link_3d=item["link_3d"],
+                    shoot_count=item["shoot_count"],
                     creat_time=item["creat_time"]
                 )
                 self.session.add(sql)
                 self.session.commit()
             except Exception as e:
                 self.session.rollback()
+                # logging.error("DBPipeline_error：", e)
                 print("DBPipeline_error：", e)
         else:
             return item
